requirements.txt
# ComfyUI 图片格式转换节点依赖
# 注意：这些依赖通常已经包含在ComfyUI环境中

torch>=1.12.0
numpy>=1.21.0
Pillow>=8.0.0
